requirement.txt
flask
sentence-transformers==2.2.2
langchain
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
pinecone[grpc]
pypdf
python-dotenv
